requirements.txt
streamlit
assemblyai
google-cloud-texttospeech
python-dotenv
st_audiorec
noisereduce
librosa
soundfile
requests
